Add AtB bus component
* Add AtbStopSensor, exposes data on the next bus from a given stop
* Add AtbDepartureSensor, exposes data on the 5 next departures from a given stop

The AtbStopSensor stores all available data from the API in it's attributes.
The AtbDepartureSensor acts as a relay, and are updated it's parent AtbStopSensor is updated, in order to minimize polling and HTTP requests.

The stop sensor can be used to quickly display the next bus from a stop, and the departure sensors can be used with a popup-card view to display additional information when clicked.

The entity_id of the departure sensor is <parent stop sensor>_<0-4>.